requirements.txt
fitz==0.0.1.dev2
gradio==5.3.0
moviepy==1.0.3
openai==1.52.0
pydantic==2.9.2
pydub==0.25.1
python-dotenv==1.0.1
